% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbgraph.R
\name{bbgraph}
\alias{bbgraph}
\title{Creates a graph of the basketball data}
\usage{
bbgraph(data, team = NULL)
}
\arguments{
\item{data}{tibble}

\item{team}{character}
}
\value{
ggplot2
}
\description{
Give a better description of what the graph is as soon as you decide the
final product of what the graph will look like.
}
