% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_filter.R
\name{team_filter}
\alias{team_filter}
\title{Filters the kenpom data by given team.}
\usage{
team_filter(data, team)
}
\arguments{
\item{data}{A tibble, preferably the kenpom data}

\item{team}{A string, the team to filter for}
}
\value{
tibble
}
\description{
Asked in the homework as part of Question 3d. Uses the filter function
to filter for a given team whether they were the home or visiting team.
}
\examples{
team_filter(get_cbb_data(), "Southern Utah")
team_filter(get_cbb_data(), "Northern Arizona")


}
