% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_list.R
\name{team_list}
\alias{team_list}
\title{Create a List of Teams}
\usage{
team_list(data)
}
\arguments{
\item{data}{Tibble of the Ken Pomeroy data.}
}
\value{
A tibble with one column of distinct teams
}
\description{
Takes the Ken Pomeroy basketball data and creates a list of distinct teams featured
in the data set. Used for the Team drop down selection in the Shiny app.
}
