% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_record_2.R
\name{team_record_2}
\alias{team_record_2}
\title{Creates a tibble of Team Stats}
\usage{
team_record_2(data)
}
\description{
This code does not run when the Shiny App is run. It takes too long. The
output of this function is saved in data/team_records.csv in the package.
}
