% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_graph.R
\name{time_series_graph}
\alias{time_series_graph}
\title{Creates a ggplot of Team and Opponent scores over time}
\usage{
time_series_graph(data, team, smooth = 0)
}
\arguments{
\item{data}{kenpom data set}

\item{team}{character string of team name}
}
\value{
ggplot2
}
\description{
Creates a ggplot of Team and Opponent scores over time
}
