% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violingraph.R
\name{violingraph}
\alias{violingraph}
\title{Builds a plot of Score Difference distribution as a violin graph}
\usage{
violingraph(data, team)
}
\arguments{
\item{data}{A tibble of the kenpom data}

\item{team}{A character string of the team name}
}
\value{
ggplot2
}
\description{
Works with the devinwcbb Shiny app to display a violin distribution of score
differences between the selected team and their opponent. A positive score
represents a win and a negative score represents a loss.
}
