% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_filter.R
\name{year_filter}
\alias{year_filter}
\title{Filters a tibble with a 'date' column for given year}
\usage{
year_filter(data, year = 2022)
}
\arguments{
\item{data}{A tibble, preferable the kenpom data}

\item{year}{An integer value representing the desired year}
}
\value{
tibble
}
\description{
The homework asked to do this as part of question 3d. Right now this
only works for year values of 2021 and 2022.
}
\examples{
year_filter(get_cbb_data(), 2021)
}
