% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_teams_records.R
\name{all_teams_records}
\alias{all_teams_records}
\title{Builds a tibble of all college men's basketball teams win records.}
\usage{
all_teams_records(data)
}
\arguments{
\item{data}{A tibble, preferably the kenpom data}
}
\value{
tibble
}
\description{
Asked in the homework as part of Question 3e. Builds a column of unique
team names. Using the 'team_win_record' function with the vector output,
builds a tibble with the columns 'team', 'wins', 'total_games', and
'win_percentage'. Sorts the final output by a weighted win percentage.
}
