% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbgraph.R
\name{bbgraph}
\alias{bbgraph}
\title{Scatter Plot of Basketball Game Scores}
\usage{
bbgraph(data, team = NULL)
}
\arguments{
\item{data}{A tibble of the Ken Pomeroy basketball data}

\item{team}{Character value of the desired team.}
}
\value{
A geom_point plot.
}
\description{
For a filtered men's college basketball team builds a scatter plot where
opponent score is on the x-axis, and team score is on the y-axis. A y = x line
is also shown to show which games are close or not
}
