% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_list.R
\name{conf_list}
\alias{conf_list}
\title{Creates a list of conferences}
\usage{
conf_list(conf_data)
}
\arguments{
\item{data}{A tibble of the Ken Pom data}
}
\value{
tibble
}
\description{
Takes the basketball data and creates a list of distinct conferences featured
in the dataset.
}
