% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_filter.R
\name{team_filter}
\alias{team_filter}
\title{Filter Ken Pomeroy data}
\usage{
team_filter(data, team)
}
\arguments{
\item{data}{Tibble of the Ken Pomeroy data}

\item{team}{Character of the desired string}
}
\value{
A tibble subset of the Ken Pomeroy data
}
\description{
Takes the Ken Pomeroy data and filters for all games that have a given team.
}
\examples{
team_filter(get_cbb_data(), "Southern Utah")
team_filter(get_cbb_data(), "Northern Arizona")


}
