% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_list.R
\name{team_list}
\alias{team_list}
\title{Creates a list of teams}
\usage{
team_list(data)
}
\arguments{
\item{data}{A tibble of the Ken Pom data}
}
\value{
tibble
}
\description{
Takes the basketball data and creates a list of distinct teams featured
in the dataset.
}
