% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_graph.R
\name{time_series_graph}
\alias{time_series_graph}
\title{Time Series Graph}
\usage{
time_series_graph(data, team, smooth)
}
\arguments{
\item{data}{Tibble of Ken Pomeroy data}

\item{team}{Character of desired string}

\item{smooth}{Binary Use smoother.}
}
\value{
A ggplot geom_line or geom_smooth
}
\description{
For a given basketball team, creates a line graph with date on the x-axis and
score on the y-axis. Has two lines for team score and opponent score. Includes
an option to use regular lines or a smoother.
}
\examples{

## Regular line plot for Duke
time_series_graph(get_cbb_data(), "Duke")

## Smoothed line plot for Duke
time_series_graph(get_cbb_data(), "Duke", TRUE)

}
